src/requirements.txt
face_recognition==1.3.0
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.2
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.4.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
